+++ golang-backend/internal/handlers/tasktimetrackinghandler/dto.go
@@ -4,13 +4,13 @@ import "github.com/aportela/doneo/internal/handlers/userhandler"
 
 type addRequest struct {
 	Summary      string `json:"summary"`
-	TotalSeconds uint   `json:"totalSeconds"`
+	TotalSeconds uint64 `json:"totalSeconds"`
 }
 
 type updateRequest struct {
 	ID           string `json:"id"`
 	Summary      string `json:"summary"`
-	TotalSeconds uint   `json:"totalSeconds"`
+	TotalSeconds uint64 `json:"totalSeconds"`
 }
 
 type TaskTimeTrakingResponse struct {
@@ -18,7 +18,7 @@ type TaskTimeTrakingResponse struct {
 	CreatedBy    userhandler.UserBaseResponse `json:"createdBy"`
 	CreatedAt    int64                        `json:"createdAt"`
 	Summary      string                       `json:"summary"`
-	TotalSeconds uint                         `json:"totalSeconds"`
+	TotalSeconds uint64                       `json:"totalSeconds"`
 }
 
 type searchResponse struct {
